rename router.py to fastapi.py
