+++ spylib/oauth/__init__.py
@@ -1,2 +1,2 @@
-from .router import init_oauth_router  # noqa: F401
+from .fastapi import init_oauth_router  # noqa: F401
 from .tokens import OfflineToken, OnlineToken  # noqa: F401
